requirements.txt
Django>=4.2
djangorestframework>=3.14
drf-yasg>=1.21
psycopg2-binary>=2.9
gunicorn>=21.2
